hw3/requirements.txt
absl-py==2.2.2
addict==2.4.0
basicsr==1.4.2
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
facexlib==0.3.0
filelock==3.18.0
filterpy==1.4.5
fonttools==4.58.0
fsspec==2025.3.2
future==1.0.0
gfpgan==1.3.8
grpcio==1.71.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
lmdb==1.6.2
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
pip==25.1.1
platformdirs==4.3.8
protobuf==6.30.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
realesrgan==0.3.0
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.3
setuptools==75.6.0
six==1.17.0
sympy==1.14.0
tb-nightly==2.20.0a20250514
tensorboard-data-server==0.7.2
tifffile==2025.5.10
tomli==2.2.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
yapf==0.43.0
